Refactor: reorganize source code into src/ directory structure
- Move source files to organized subdirectories (haplotag, phase, modcall, somatic_haplotag, shared)
- Update Makefile compilation rules for new structure
- Remove executable from git tracking
- Improve code organization and maintainability
