+++ .gitignore
@@ -35,7 +35,7 @@ out/
 .DS_Store
 Thumbs.db
  
-longphase 
+longphase-s 
 htslib/*
 jemalloc/*
 configure
