+++ src/haplotag/HaplotagLogging.cpp
