+++ src/haplotag/HaplotagLogging.h
