+++ src/haplotag/HaplotagParsingBam.cpp
