+++ src/haplotag/HaplotagStrategy.cpp
