+++ src/haplotag/HaplotagStrategy.h
