+++ src/haplotag/HaplotagVcfParser.cpp
