+++ src/haplotag/HaplotagVcfParser.h
