+++ src/modcall/ModCallProcess.h
@@ -1,6 +1,6 @@
 #ifndef MODCALLPROCESS_H
 #define MODCALLPROCESS_H
-#include "Util.h"
+#include "../shared/Util.h"
 
 
 struct ModCallParameters
