+++ src/phase/Phasing.cpp
@@ -1,6 +1,6 @@
 #include "Phasing.h"
 #include "PhasingProcess.h"
-#include "Util.h"
+#include "../shared/Util.h"
 #include <getopt.h>
