+++ src/shared/ArgumentManager.cpp
