+++ src/shared/ArgumentManager.h
