+++ src/somatic_haplotag/SomaticBenchmark.cpp
