+++ src/somatic_haplotag/SomaticHaplotag.cpp
