+++ src/somatic_haplotag/SomaticHaplotagProcess.cpp
