+++ src/somatic_haplotag/SomaticVarCaller.cpp
